pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.5.2",
    "flask>=3.1.0",
    "gunicorn>=23.0.0",
    "psutil>=7.0.0",
    "pynacl>=1.5.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "waitress>=3.0.2",
]
